+++ app/ai-in-aviation/page.tsx
@@ -1,10 +1,10 @@
-import AiSolution from "../components/aiSolution";
-import BookConsultation from "../components/bookConsultation";
-import ClientsSection from "../components/clientSection";
-import ConsultationBanner from "../components/consultationBanner";
-import HeroSection from "../components/heroSection";
-import AviationOverview from "../components/industryOverview";
-import KeyBenefits from "../components/keyBenefits";
+import AiSolution from "../../components/aiSolution";
+import BookConsultation from "../../components/bookConsultation";
+import ClientsSection from "../../components/clientSection";
+import ConsultationBanner from "../../components/consultationBanner";
+import HeroSection from "../../components/heroSection";
+import AviationOverview from "../../components/industryOverview";
+import KeyBenefits from "../../components/keyBenefits";
 import {
   aiSolutions,
   aviationIndustryOverview,
@@ -12,8 +12,8 @@ import {
   keyBenefits,
   technologiesSection,
   consultationBanner,
-} from "../lib/ai-in-aviation-data";
-import { clientData } from "../lib/clientData";
+} from "../../lib/ai-in-aviation-data";
+import { clientData } from "../../lib/clientData";
 
 const AiInAviation = () => {
   return (
